# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "The API for the PippHo IoT project"
  version: "1.0.0"
  title: "pipp-ho"
host: "virtserver.swaggerhub.com"
basePath: "/Fanatic17/PippHo/1.0.0"
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /command/{deviceId}/{sensorName}/{value}:
    get:
      tags:
      - "Command"
      description: "command"
      operationId: "getComnmands"
      parameters:
      - name: "deviceId"
        in: "path"
        required: true
        type: "string"
      - name: "sensorName"
        in: "path"
        required: true
        type: "string"
      - name: "value"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Command"
definitions:
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
        example: "everything is ok"
    example:
      code: 0
      message: "everything is ok"